Fixed time interval not appearing after first delivery

+++ Assets/Scripts/GameManager.cs
@@ -84,7 +84,7 @@ public class GameManager : MonoBehaviour
 
     public void AddTime(int modifierMultiplier)
     {
-        StartCoroutine(UIManager.instance.AddTimeUI((float)extraTimeModifier*modifierMultiplier));
+        StartCoroutine(UIManager.instance.AddTimeUI(extraTimeModifier*modifierMultiplier));
         extraTime += extraTimeModifier*modifierMultiplier;
     }
